ai_resume_anlayzer/requirements.txt
streamlit
pdfplumber
python-docx
scikit-learn
ollama

